Added simple visualiser to show orbits/satellites in visualiser.py
